✨feat : 세미나 4주차 과제 (사용장 CR 실습_수정)
