+++ src/main/java/com/project/likelion13thbe/domain/member/service/query/MemberQueryService.java
@@ -1,0 +1,4 @@
+package com.project.likelion13thbe.domain.member.service.query;
+
+public interface MemberQueryService {
+}
